Cleanup after migration
