+++ src/main/kotlin/no/nav/kafka/dialog/nais/NaisDSL.kt
@@ -3,9 +3,7 @@ package no.nav.kafka.dialog
 import mu.KotlinLogging
 import no.nav.kafka.dialog.gui.filesHandler
 import no.nav.kafka.dialog.metrics.Prometheus
-import no.nav.kafka.dialog.salesforce.DefaultAccessTokenHandler
-import no.nav.sf.pdl.kafka.salesforce.NewAccessTokenHandler
-import no.nav.sf.pubsub.token.MigratingAccessTokenHandler
+import no.nav.sf.pdl.kafka.salesforce.DefaultAccessTokenHandler
 import org.http4k.core.HttpHandler
 import org.http4k.core.Method
 import org.http4k.core.Response
@@ -36,10 +34,8 @@ fun naisAPI(): HttpHandler =
                 Response(Status.INTERNAL_SERVER_ERROR)
             }
         },
-        "/internal/testAccess/old" bind Method.GET to testAccessHandlerOld,
         "/internal/testAccess/new" bind Method.GET to testAccessHandlerNew,
         "/internal/testAccess/validation" bind Method.GET to testAccessHandlerValidation,
-        "/internal/testAccess/migration" bind Method.GET to testAccessHandlerMigration,
         "/internal/files" bind Method.GET to filesHandler(File("/tmp/files")),
         "/internal/files/{path:.*}" bind Method.GET to filesHandler(File("/tmp/files")),
         // "/internal/gui" bind Method.GET to Gui.guiHandler
@@ -71,25 +67,14 @@ object ShutdownHook {
     fun isActive() = shutdownhookActive
 }
 
-private val testAccessHandlerOld: HttpHandler = {
-    val defaultAccessTokenHandler = DefaultAccessTokenHandler()
-    Response(OK).body("Test access (old) successful: " + defaultAccessTokenHandler.testAccess())
-}
-
 private val testAccessHandlerNew: HttpHandler = {
-    val newAccessTokenHandler = NewAccessTokenHandler()
+    val newAccessTokenHandler = DefaultAccessTokenHandler()
     Response(OK).body("Test access (new) successful: " + newAccessTokenHandler.testAccess())
 }
 
 private val testAccessHandlerValidation: HttpHandler = {
-    val newAccessTokenHandlerAgainstValidation = NewAccessTokenHandler(sfClientId = env(secret_SF_VALIDATION_CLIENT_ID))
+    val newAccessTokenHandlerAgainstValidation = DefaultAccessTokenHandler(sfClientId = env(secret_SF_VALIDATION_CLIENT_ID))
     Response(OK).body("Test access (validation) successful: " + newAccessTokenHandlerAgainstValidation.testAccess())
 }
 
 val currentTimeStamp: String get() = LocalDateTime.now().format(DateTimeFormatter.ISO_DATE_TIME)
-
-private val testAccessHandlerMigration: HttpHandler = {
-    val migrationTokenHandler =
-        MigratingAccessTokenHandler(old = DefaultAccessTokenHandler(), new = NewAccessTokenHandler())
-    Response(OK).body("$currentTimeStamp\nTest access (migration) result: " + migrationTokenHandler.testAccess())
-}
